2
now the memory at heads up all the weight you did regardless of number of repetitions repetition stay at 10 which is annoying and it averages it so I did like the full set of incline deck push-up whatever it's called and then added them up and divided by the number of sets which was four so I did like 56.5 which is not what I wanted. I wanted to have the latest set so if it does if I do like 40 kg on the third set and 50 kg on the fourth set, I wanted to remember the fourth set the last set so I gotta walk on that.