6
The heart rate right now in the profile view is accurate but there's problems with tapping on the widget and the heart rate view which is View is an issue. Also the heart rate and the active workout view is also an accurate just like the last I have to fix it probably in a similar way to how I fix the widget in the profile view just something to keep in mind for the next session.