17?
The countdown logic is finally fixed. I did this by creating a separate file and connecting it via data manager and updating the main file to make sure it all sinks up the countdown should work when the app is in the background and obviously one switching tabs so I'm very glad this is over.