19
This is a problematic one. I fixed swiping between each screen by removing navigation view from each and every tab now it's native it works it's pretty smooth. I have to work a little bit on performance. It broke a lot of stuff so now I have to fix this from what I've seen it. It broke the history tab which is not interact anymore and the template detail View which is seems like the workout View is also broken. I'll have to fix this.